refactor: extract card_table_panel grid renderers into table_renderers.py
Move the four wx.grid cell renderers (_ManaIconCellRenderer,
_EllipsisStringRenderer, _InlineSymbolStringRenderer, _ActionCellRenderer),
their painting helpers (_scaled_bitmap, _split_inline_symbols, _font_height,
_paint_row_background), the shared bitmap cache, and the render constants out of
table_view.py into a sibling table_renderers.py module. DeckTableView stays in
table_view.py and imports the renderers from the new module, preserving its
public import path. Pure behavior-preserving move; no rendering logic changed.

Closes #538
